Homework1copia

+++ Homework1Copia/AreaCircunferencia2.java
@@ -3,7 +3,7 @@ import javax.swing.JOptionPane;
 public class AreaCircunferencia2{
   public static void main(String[] args) {
 
-    double r = Integer.parseInt(JOptionPane.showInputDialog(null, "Ingrese el radio de la circunferencia"));
+    double r = Double.parseDouble(JOptionPane.showInputDialog(null, "Ingrese el radio de la circunferencia"));
     double pi;
     pi = 3.1416;
     double area;
