+++ Homework1Copia/AreaFigura2.java
@@ -2,8 +2,8 @@ import javax.swing.JOptionPane;
 
 public class AreaFigura2{
   public static void main(String[] args) {
-    double r = Integer.parseInt(JOptionPane.showInputDialog(null,"Ingrese el valor de R para la figura"));
-    double h = Integer.parseInt(JOptionPane.showInputDialog(null,"Ingrese el valor de H para la figura"));
+    double r = Double.parseDouble(JOptionPane.showInputDialog(null,"Ingrese el valor de R para la figura"));
+    double h = Double.parseDouble(JOptionPane.showInputDialog(null,"Ingrese el valor de H para la figura"));
     double pi;
     pi = 3.1416;
     double c;
