+++ Homework1Copia/AreaTerreno2.java
@@ -2,9 +2,9 @@ import javax.swing.JOptionPane;
 
 public class AreaTerreno2{
   public static void main(String[] args) {
-    double b = Integer.parseInt(JOptionPane.showInputDialog(null,"Ingrese la medida del lado B del terreno"));
-    double a = Integer.parseInt(JOptionPane.showInputDialog(null,"Ingrese la medida del lado A del terreno"));
-    double c = Integer.parseInt(JOptionPane.showInputDialog(null,"Ingrese la medida del lado C del terreno"));
+    double b = Double.parseDouble(JOptionPane.showInputDialog(null,"Ingrese la medida del lado B del terreno"));
+    double a = Double.parseDouble(JOptionPane.showInputDialog(null,"Ingrese la medida del lado A del terreno"));
+    double c = Double.parseDouble(JOptionPane.showInputDialog(null,"Ingrese la medida del lado C del terreno"));
     double at;
     double ar;
     double area;
