+++ Homework1Copia/ProductorLeche2.java
@@ -3,7 +3,7 @@ import javax.swing.JOptionPane;
 public class ProductorLeche2{
   public static void main(String[] args) {
     int l = Integer.parseInt(JOptionPane.showInputDialog(null,"Ingrese los litros de leche producidos"));
-    double pg = Integer.parseInt(JOptionPane.showInputDialog(null,"Ingrese el precio del galon de leche"));
+    double pg = Double.parseDouble(JOptionPane.showInputDialog(null,"Ingrese el precio del galon de leche"));
     double tg;
     tg = l / 3.785;
     double ga;
